Fix file descriptor sharing after re-exec
Previously if you ran `commands | grep foo`, it would never output
anything. This seems to be because the file descriptor isn't shared with
the exec'd process unless it is open https://stackoverflow.com/a/55502228/902968

This does the simplest version of opening by also printing it's
re-execuing as root, which explains the password prompt you see.